+++ chocolatey-visualstudio.extension/extensions/Install-VisualStudio.ps1
@@ -37,7 +37,9 @@ Install-ChocolateyPackage
       [string] $Checksum,
       [string] $ChecksumType,
       [ValidateSet('MsiVS2015OrEarlier', 'WillowVS2017OrLater')] [string] $InstallerTechnology,
-      [string] $ProgramsAndFeaturesDisplayName = $ApplicationName
+      [string] $ProgramsAndFeaturesDisplayName = $ApplicationName,
+      [string] $VisualStudioYear,
+      [string] $Product
     )
     if ($Env:ChocolateyPackageDebug -ne $null)
     {
@@ -45,21 +47,31 @@ Install-ChocolateyPackage
         $DebugPreference = 'Continue'
         Write-Warning "VerbosePreference and DebugPreference set to Continue due to the presence of ChocolateyPackageDebug environment variable"
     }
-    Write-Debug "Running 'Install-VisualStudio' for $PackageName with ApplicationName:'$ApplicationName' Url:'$Url' Checksum:$Checksum ChecksumType:$ChecksumType InstallerTechnology:'$InstallerTechnology' ProgramsAndFeaturesDisplayName:'$ProgramsAndFeaturesDisplayName'";
+    Write-Debug "Running 'Install-VisualStudio' for $PackageName with ApplicationName:'$ApplicationName' Url:'$Url' Checksum:$Checksum ChecksumType:$ChecksumType InstallerTechnology:'$InstallerTechnology' ProgramsAndFeaturesDisplayName:'$ProgramsAndFeaturesDisplayName' VisualStudioYear:'$VisualStudioYear' Product:'$Product'";
 
     $assumeNewVS2017Installer = $InstallerTechnology -eq 'WillowVS2017OrLater'
 
-    $uninstallKey = Get-VSUninstallRegistryKey -ApplicationName $ProgramsAndFeaturesDisplayName
-    $count = ($uninstallKey | Measure-Object).Count
-    if ($count -gt 0)
+    if ($assumeNewVS2017Installer)
     {
-        if ($assumeNewVS2017Installer)
+        # there is a single Programs and Features entry for all products, so its presence is not enough
+        if ($VisualStudioYear -ne '' -and $Product -ne '')
         {
-            # TODO: implement detection of products installed by Willow
-            # (there is a single Programs and Features entry for all products, so its presence is not enough)
-            Write-Debug "Programs and Features entry '$ProgramsAndFeaturesDisplayName' detected, this or other product from this family is installed"
+            $prodRef = Get-VSProductReference -VisualStudioYear $VisualStudioYear -Product $Product
+            $products = Get-WillowInstalledProducts | Where-Object { $_ -ne $null -and $_.channelId -eq $prodRef.ChannelId -and $_.productId -eq $prodRef.ProductId }
+            $productsCount = ($products | Measure-Object).Count
+            Write-Verbose ("Found {0} installed Visual Studio product(s) with ChannelId = {1} and ProductId = {2}" -f $productsCount, $prodRef.ChannelId, $prodRef.ProductId)
+            if ($productsCount -gt 0)
+            {
+                Write-Warning "$ApplicationName is already installed. Please use the Visual Studio Installer to modify or repair it."
+                return
+            }
         }
-        else
+    }
+    else
+    {
+        $uninstallKey = Get-VSUninstallRegistryKey -ApplicationName $ProgramsAndFeaturesDisplayName
+        $count = ($uninstallKey | Measure-Object).Count
+        if ($count -gt 0)
         {
             Write-Warning "$ApplicationName is already installed. Please use Programs and Features in the Control Panel to modify or repair it."
             return